ISSUE #3626 - extract commentReply
